changed frame injection impl
previous implementation inserted two statements into the stmt vector
owned by the block object. this became a problem when the function
returned a value, as the footer statement was not being executed.

the new impl instead takes a different approach of nesting the function
block inside another function block which captures the return statement,
and thereby ensures that the footer statement is always executed.